WIP: Unified high-level key classes
