--- src/Cryptography/OpenPgp/IPgpKey.cs
@@ -1,8 +1,0 @@
-﻿namespace Springburg.Cryptography.OpenPgp
-{
-    interface IPgpKey
-    {
-        long KeyId { get; }
-        bool IsMasterKey { get; }
-    }
-}
