+++ src/Cryptography/OpenPgp/Keys/S2kParameters.cs
@@ -3,7 +3,7 @@
 
 namespace Springburg.Cryptography.OpenPgp.Keys
 {
-    class S2kParameters
+    public class S2kParameters
     {
         public S2kUsageTag UsageTag { get; set; } = S2kUsageTag.Sha1;
         public PgpSymmetricKeyAlgorithm EncryptionAlgorithm { get; set; } = PgpSymmetricKeyAlgorithm.Aes128;
