+++ test/PGPClearSignedSignatureTest.cs
@@ -254,7 +254,7 @@ namespace Org.BouncyCastle.Bcpg.OpenPgp.Tests
             using (var messageGenerator = new PgpMessageGenerator(new ArmoredPacketWriter(bOut)))
             using (var signedGenerator = messageGenerator.CreateSigned(PgpSignatureType.CanonicalTextDocument, pgpPrivKey, PgpHashAlgorithm.Sha256))
             {
-                signedGenerator.HashedAttributes.SetSignerUserId(false, pgpSecKey.PublicKey.GetUserIds().First().UserId);
+                signedGenerator.HashedAttributes.SetSignerUserId(false, pgpSecKey.GetUserIds().First().UserId);
                 using (var literalStream = signedGenerator.CreateLiteral(PgpDataFormat.Text, "", DateTime.MinValue))
                 {
                     literalStream.Write(Encoding.UTF8.GetBytes(message));
